requirements.txt
asgiref==3.5.2
attrs==22.2.0
az-iranian-bank-gateways==1.8.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-jalali==6.0.1
django-jalali-date==1.0.1
django-mysql==4.8.0
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
jdatetime==4.1.0
lxml==4.9.2
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
suds==1.1.2
tzdata==2022.7
urllib3==1.26.13
zeep==4.2.1
